requirements.txt
torch
torchvision
opencv-python
onnx
onnxruntime
requests
tqdm
numpy
Pillow
